get_motion_kernel
