+++ kornia/geometry/transform/pyramid.py
@@ -6,7 +6,8 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
-from kornia.filters import gaussian_blur2d, filter2d
+from kornia.filters.gaussian import gaussian_blur2d
+from kornia.filters.filter import filter2d
 
 __all__ = ["PyrDown", "PyrUp", "ScalePyramid", "pyrdown", "pyrup", "build_pyramid"]
